% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\docType{data}
\name{F255744}
\alias{F255744}
\title{F255744 Mascot Search results}
\description{
F255744 Mascot Search results
}
\examples{
class(F255744)
hist(F255744$RTINSECONDS)
hist(F255744$RTINSECONDS[F255744$pep_score>20])
}
\author{
Pascal Egloff \email{p.egloff@imm.uzh.ch}
}
\keyword{data}
