% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{getFC}
\alias{getFC}
\title{Read FlyCodes (FC)}
\usage{
getFC(pattern = "^GS[ASTNQDEFVLYWGP]{7}(WR|WLTVR|WQEGGR|WLR|WQSR)$")
}
\arguments{
\item{pattern}{a regular expression FlyCode pattern}
}
\value{
a data.frame of FlyCodes
}
\description{
Read FlyCodes (FC)
}
\examples{
FC <- getFC()
dim(FC)

}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2015
}
