% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{getNB}
\alias{getNB}
\title{Read NanoBodies (NBs)}
\usage{
getNB()
}
\value{
a data.frame of NBs
}
\description{
Read NanoBodies (NBs)
}
\examples{
NB <- getNB()
dim(NB)
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2015, 2018
}
