% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{nanobodyFlycodeLinking.as.fasta}
\alias{nanobodyFlycodeLinking.as.fasta}
\title{write FASTA}
\usage{
nanobodyFlycodeLinking.as.fasta(x, ...)
}
\arguments{
\item{x}{a \code{nanobodyFlycodeLinking} class computed by \code{\link{runNGSAnalysis}}.}

\item{...}{just passed}
}
\value{
sprintf stream
}
\description{
write FASTA
}
\examples{
\dontrun{
FF <- c('~/__projects/2018/20181030-p1875/NL7-1.extendedFrags_uniqNB2FC.txt', 
'~/__projects/2018/20181030-p1875/NL7-2.extendedFrags_uniqNB2FC.txt')

X <- do.call('rbind', lapply(FF, read.table, header=T, sep='\\t'))
cat(uniqNB2FC.as.fasta(X), file=tempfile(fileext=".fasta"))
}
}
\author{
Lennart Opitz, Christian Panse 2018
}
