% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{nanobodyFlycodeLinking.summary}
\alias{nanobodyFlycodeLinking.summary}
\title{Object Summaries of S3 class \code{nanobodyFlycodeLinking}}
\usage{
nanobodyFlycodeLinking.summary(object)
}
\arguments{
\item{object}{a \code{nanobodyFlycodeLinking} class computed
by \code{\link{runNGSAnalysis}}.}
}
\value{
a data.frame object
}
\description{
Object Summaries of S3 class \code{nanobodyFlycodeLinking}
}
\examples{
f <- system.file("extdata/nanobodyFlycodeLinkage.RData", package="NestLink")
load(f)
summary(nanobodyFlycodeLinkage.sample)
}
