% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{NB.unambiguous}
\alias{NB.unambiguous}
\title{Determine unambiguous NBs}
\usage{
NB.unambiguous(x)
}
\arguments{
\item{x}{a \code{data.frame} containing a column peptide}
}
\value{
a data.frame 
a \code{data.frame} of unambiguously assignable peptides 
(those, which occur only on one nanobody)
}
\description{
Determine unambiguous NBs
}
