% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{plot_in_silico_LCMS_map}
\alias{plot_in_silico_LCMS_map}
\title{plot a LC-MS map}
\usage{
plot_in_silico_LCMS_map(peptides, ...)
}
\arguments{
\item{peptides}{a vector of pepitdes.}

\item{...}{pass through the plot method.}
}
\value{
gplots::hist2d a gplot 2d histogram
}
\description{
plot a LC-MS map
}
\examples{
set.seed(1)
plot_in_silico_LCMS_map(FlyCodes <- replicate(10, compose_GPx10R()))
}
\author{
Christian Panse
}
