% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_NGS.r
\name{runNGSAnalysis}
\alias{runNGSAnalysis}
\title{run NGS workflow}
\usage{
runNGSAnalysis(file, param)
}
\arguments{
\item{file}{- sequence file path}

\item{param}{- list of input parameters}
}
\description{
run NGS workflow
}
