% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions_NGS.r
\name{TwoPatternReadFilter}
\alias{TwoPatternReadFilter}
\title{filter input sequences for two patterns}
\usage{
TwoPatternReadFilter(reads, leftPattern, rightPattern, maxMismatch,
  previousPatternPos = NULL)
}
\arguments{
\item{reads}{}

\item{leftPattern}{}

\item{rightPattern}{}

\item{maxMismatch}{}

\item{previousPatternPos}{}
}
\value{
list object
}
\description{
filter input sequences for two patterns
}
