% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose_GPx10R}
\alias{compose_GPx10R}
\title{Compose a peptide with a defined AA sequence}
\usage{
compose_GPx10R(aa_pool1, aa_pool2)
}
\arguments{
\item{aa_pool1}{AA distributen.}

\item{aa_pool2}{AA distributen.}
}
\value{
a AA sequence
}
\description{
Compose a peptide with a defined AA sequence
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2015
}
