% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compose.R
\name{compose_GPx10R}
\alias{compose_GPx10R}
\title{compose a peptide with a defined AA sequence}
\usage{
compose_GPx10R(aa_pool1, aa_pool2)
}
\arguments{
\item{pool}{AA distributen.}

\item{pool}{AA distributen.}
}
\value{
a AA sequence
}
\description{
compose a peptide with a defined AA sequence
}

