% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{getNB}
\alias{getNB}
\title{Read NanoBodies (NBs)}
\source{
\itemize{
\item{\url{https://fgcz-gstore.uzh.ch/projects/p1644/analysis_20170609_o3040/p1644o3482-4_S4.extendedFrags_uniqNB2FC.txt}}
\item{\url{https://fgcz-gstore.uzh.ch/projects/p1644/analysis_20170609_o3040/p1644o3482-5_S5.extendedFrags_uniqNB2FC.txt}}
}
}
\usage{
getNB()
}
\value{
a \code{data.frame} object of NBs
}
\description{
A wrapper function for reading the
flycodes using ExperimentHub.
The files are used for demonstrating the detectability of the AA
sequences. The wrapper functions are extended by columns
\code{\link[protViz]{ssrc}} prediction
and the \code{\link[protViz]{parentIonMass}}.
The column ESP_Prediction was generated by using the service from
\url{https://genepattern.broadinstitute.org}.
}
\examples{
NB <- getNB()
dim(NB)
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2015, 2018
}
