% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{nanobodyFlycodeLinking.as.fasta}
\alias{nanobodyFlycodeLinking.as.fasta}
\title{Write FASTA}
\usage{
nanobodyFlycodeLinking.as.fasta(x, file = NULL, ...)
}
\arguments{
\item{x}{a \code{nanobodyFlycodeLinking} S3 object computed by
\code{\link{runNGSAnalysis}}.}

\item{file}{a filename}

\item{...}{just passed}
}
\value{
sprintf stream
}
\description{
Write FASTA
}
\examples{
library(ExperimentHub)
eh <- ExperimentHub()
f <- query(eh, c("NestLink", "nanobodyFlycodeLinkage.RData"))[[1]]
load(f)
summary(nanobodyFlycodeLinkage.sample)
nanobodyFlycodeLinking.as.fasta(nanobodyFlycodeLinkage.sample)
}
\author{
Lennart Opitz, Christian Panse 2018
}
