% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\docType{data}
\name{PGexport}
\alias{PGexport}
\title{PGexport results}
\description{
PGexport results
}
\examples{
filename <- system.file("extdata/PGexport2_normalizedAgainstSBstandards_Peptides.csv",
  package = "NestLink")
P <- read.csv(filename,
  header = TRUE, sep=';')
}
\author{
Pascal Egloff \email{p.egloff@imm.uzh.ch}
}
\keyword{data}
