% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{plot_in_silico_LCMS_map}
\alias{plot_in_silico_LCMS_map}
\title{plot a LC-MS map of a given set of amino acid sequences}
\usage{
plot_in_silico_LCMS_map(peptides, ...)
}
\arguments{
\item{peptides}{a vector of pepitdes.}

\item{...}{pass through the plot method.}
}
\value{
gplots::hist2d a gplot 2d histogram
}
\description{
plot a LC-MS map of a given set of amino acid sequences
}
\details{
TODO(cp): consider using hexbin using ggplot2 ggplot facet_wrap aes geom_point
}
\examples{
set.seed(1)
par(mfrow=c(2,1));
rv <- plot_in_silico_LCMS_map(FlyCodes <- replicate(10000, compose_GPGx8cTerm()))
}
\author{
Christian Panse
}
