% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions_NGS.r
\name{twoPatternReadFilter}
\alias{twoPatternReadFilter}
\title{filter input sequences for two patterns}
\usage{
twoPatternReadFilter(reads, leftPattern, rightPattern, maxMismatch,
  prevPatternPos = NULL)
}
\arguments{
\item{reads}{}

\item{leftPattern}{}

\item{rightPattern}{}

\item{maxMismatch}{}

\item{prevPatternPos}{}
}
\value{
list object
}
\description{
filter input sequences for two patterns
}
\examples{
reads <- DNAStringSet(c('ACTGGGTTT','ACCCTGGGTTT'))
leftPattern <- 'CT'
rightPattern <- 'TTT'
maxMismatch <- 0
twoPatternReadFilter(reads, leftPattern, rightPattern, maxMismatch)
}
