% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\docType{data}
\name{WU160118}
\alias{WU160118}
\title{WU160118 Mascot Search results}
\description{
WU160118 Mascot Search results
}
\examples{
library(ExperimentHub)
eh <- ExperimentHub(); 
load(query(eh, c("NestLink", "WU160118.RData"))[[1]])
class(WU160118)
PATTERN <- "^GS[ASTNQDEFVLYWGP]{7}(WR|WLTVR|WQEGGR|WLR|WQSR)$"
idx <- grepl(PATTERN, WU160118$pep_seq)
WU <- WU160118[idx & WU160118$pep_score > 25,]

library(lattice)
histogram(~RTINSECONDS| datfilename, data = WU, type='count')
}
\references{
\url{https://fgcz-bfabric.uzh.ch/bfabric/userlab/show-workunit.html?id=160118}
}
\author{
Christian Panse
}
\keyword{data}
